WIP: Allow multiple EARLY_PRINTF drivers
Only one is able to be selected in any given kernel, but with this we
can have multiple drivers available to be built. We select which we
should use by using something like (for the pl011 UART driver):
options EARLY_PRINTF=uart_pl011

The console code will then set:
early_putc_t *early_putc = uart_pl011_early_putc;

Another driver can then also be defined with a different putc function
that can then be enabled by changing the config file.

To select the address for the uart to use we use SOCDEV_VA/SOCDEV_PA.
This works as we mask out the lower addresses when adding an entry to
the page table.