Make boot1.efi use shim when available.
When running with Secure Boot enabled; LoadImage() and StartImage()
UEFI functions will fail if the image to be executed is not signed
with certificate known to UEFI, even if it is correctly signed with
certificate known to the shim.  This patch makes it use Execute()
method provided by shim.  This uses the exact same mechanism the
shim itself uses to execute boot1.efi.