Enable SMP and start scheduling on multiple cores.
We use the FDT data to find the CPUs to start, limiting them to MAXCPU.
As we need the psci driver to start the cores we only start them late in
the boot when other architectures would be releasing them from a busy loop.
This can lead to increases in the time it takes to boot and may need to
be evaluated when we are trying to start on a larger number of cores.

Partially copied from a similar patch Robin Randhawa <Robin.Randhawa@ARM.com>
for bits I missed on my initial change.