Add the start of a version of loader that looks like Linux.
The kernel is loaded by the firmware as a Linux init ramdisk. The
loader then looks through the device tree for this and boots it.
It currently assumes the kernel is correctly aligned, this will
be broken in general, but works on the Dragonboard 410c where this
has been tested.